requirements.txt
# Core
pandas>=1.5.0
Pillow>=9.0.0
numpy>=1.23.0
scikit-learn>=1.1.0

# Image processing
albumentations>=1.3.0
opencv-python-headless>=4.7.0

# OCR
paddlepaddle-gpu
paddleocr>=2.7.0

# Deep Learning & KIE
torch>=2.0.0
transformers>=4.36.0
datasets>=2.14.0

# Evaluation
seqeval>=1.2.2

# Visualization
matplotlib>=3.6.0

# Dataset download
gdown>=4.7.1
